Move fs func out of mod
