+++ mod.js
@@ -2,113 +2,10 @@
 
 const { args, env} = Deno;
 import { parse } from 'https://deno.land/std@0.51.0/flags/mod.ts';
-import { ensureDir, readFileStr, writeFileStr, walk, exists } from "https://deno.land/std@0.51.0/fs/mod.ts";
 import * as logger from 'https://deno.land/std/log/mod.ts';
+import * as bsFs from './fs.js';
 
 const inputArgs = parse(args);
-
-const tempTemplate = `#!/bin/sh
-# generated by bs 'create script' 
-`;
-
-const bashProfilePath = `${env.get('HOME')}/.bash_profile`;
-
-const bsHome = `${env.get('HOME')}/.bs`;
-
-const bsBin = `${bsHome}/bin`;
-const bsBinGlobals = `${bsBin}/global`;
-const bsBinProjects = `${bsBin}/project`;
-const bsBinProject = (projectName) => `${bsBinProjects}/${projectName}`;
-
-const bsAlias = `${bsHome}/alias`;
-const bsAliasGlobals = `${bsAlias}/global`;
-const bsAliasProjects = `${bsAlias}/project`;
-const bsAliasProject = (projectName) => `${bsAliasProjects}/${projectName}`;
-
-const bsAliasMain = (aliasDir) => `${aliasDir}/main.sh`;
-
-const createAliasFilePath = (dir, name) => `${dir}/${name}.sh`
-const createAliasStr = (name, command) => `alias ${name}="${command}"`
-
-const toSourceString = (filePath) => `source ${filePath}`;
-
-async function ensureBs() {
-  await ensureDir(bsBinGlobals);
-  await ensureDir(bsBinProjects);
-  await ensureDir(bsAliasGlobals);
-  await ensureDir(bsAliasProjects);
-  return true;
-}
-
-
-async function upsertFolderMain(aliasDir) {
-  const aliases = [];
-  const mainPath = bsAliasMain(aliasDir);
-  for await (const entry of walk(aliasDir)) {
-    if (entry && entry.isFile && entry.name !== 'main.sh') {
-      const alias = await readFileStr(`${aliasDir}/${entry.name}`, { encoding: 'utf8' });
-      aliases.push(alias);
-    }
-  }
-  await writeFileStr(mainPath, aliases.join('\n'));
-  return mainPath;
-}
-
-async function getFileNames(dir, includeMain) {
-  const fileNames = [];
-  for await (const entry of walk(dir)) {
-    if (entry.isFile && (!includeMain && !entry.name.includes('main.sh'))) {
-      fileNames.push(entry.name);
-    }
-  }
-  return fileNames;
-}
-
-async function updateBashProfile(source) {
-  const org = await readFileStr(bashProfilePath, { encoding: 'utf8' });
-  if (org.includes(source)) {
-    return true;
-  }
-  return writeFileStr(bashProfilePath, `${org}\n${source}\n`);
-}
-
-async function createAliasFile(dir, name, command) {
-  const filePath = createAliasFilePath(dir, name);
-  const aliasString = createAliasStr(name, command);
-
-  await writeFileStr(filePath, aliasString);
-  return {
-    filePath,
-    aliasString,
-  };
-}
-
-async function getProjectDir(projectPathFn) {
-  const getGitRemote = Deno.run({
-    cmd: ['git', 'remote', '-v'],
-    stdout: "piped",
-    stderr: "piped",
-  });
-  const { code } = await getGitRemote.status();
-  if (code === 0) {
-    const rawOutput = await getGitRemote.output();
-    const gitRemoteString = new TextDecoder().decode(rawOutput);
-    // TODO: regEx
-    const projectName = gitRemoteString.split('\n')[0]
-      .replace('origin\t', '')
-      .replace('(fetch)', '')
-      .replace(' ', '')
-      .replace('.git', '');
-
-    const projectPath = projectPathFn(projectName);
-    await ensureDir(projectPath);
-    return projectPath;
-  } else {
-    const errorString = new TextDecoder().decode(rawError);
-    logger.error(errorString);
-    throw new Error('Unable to get the git remote Origin');
-  }
-}
 
 const commands = {
   create: {
@@ -117,7 +14,7 @@ const commands = {
         const isGlobal = (flags.g || flags.global);
         const originalCommand = flags.org || flags.o;
         const name = flags.name || flags.n;
-        const aliasDir = isGlobal ? bsAliasGlobals : await getProjectDir(bsAliasProject);
+        const aliasDir = isGlobal ? bsFs.paths.alias.globals : await bsFs.getProjectDir(bsFs.paths.alias.project);
 
         if (!name || !originalCommand) {
           // TODO: return help
@@ -126,19 +23,21 @@ const commands = {
           );
         }
 
-        await ensureBs();
+        await bsFs.ensureBs();
 
-        const { aliasPath, aliasString } = await createAliasFile(aliasDir, name, originalCommand);
-        const aliasMainPath = await upsertFolderMain(aliasDir)
-        const sourceString = toSourceString(aliasMainPath);
+        const { aliasPath, aliasString } = await bsFs.createAliasFile(
+          aliasDir, name, originalCommand
+        );
+        const aliasMainPath = await bsFs.upsertFolderMain(aliasDir)
+        const sourceString = bsFs.toSourceString(aliasMainPath);
 
         if (isGlobal) {
-          await updateBashProfile(sourceString);
+          await bsFs.updateBashProfile(sourceString);
           // TODO: this returns 404 not found... why?
           // const sourceFile = Deno.run({
           //   cmd: ['source', aliasMainPath],
           // });
-          return console.log(toSourceString(bashProfilePath));
+          return console.log(bsFs.toSourceString(bsFs.bashProfilePath));
         }
 
         return console.log(sourceString);
@@ -149,7 +48,7 @@ const commands = {
     },
     project: {
       exec: async (flags) => {
-        const projectPath = await getProjectDir(bsBinProject);
+        const projectPath = await bsFs.getProjectDir(bsFs.paths.bin.project);
         return logger.info(`Project folder at: ${projectPath}`);
       },
     },
@@ -159,9 +58,8 @@ const commands = {
         if (!name) {
           return logger.error('Name (--name, -n), is required');
         }
-        const projectPath = await getProjectDir(bsBinProject);
-        const filePath = `${projectPath}/${name}.sh`;
-        return writeFileStr(filePath, tempTemplate);
+        const projectPath = await bsFs.getProjectDir(bsFs.paths.bin.project);
+        return bsFs.createScript(name, projectPath);
       },
     },
   },
@@ -169,19 +67,19 @@ const commands = {
     alias: {
       exec: async (flags) => {
         const isGlobal = (flags.g || flags.global);
-        const aliasDir = isGlobal ? bsAliasGlobals : await getProjectDir(bsAliasProject);
-        const mainPath = bsAliasMain(aliasDir);
+        const aliasDir = isGlobal ? bsFs.paths.alias.globals : await bsFs.getProjectDir(bsFs.paths.alias.project);
+        const mainPath = bsFs.paths.alias.main(aliasDir);
         if (isGlobal) {
-          return console.log(toSourceString(bashProfilePath));
+          return console.log(bsFs.toSourceString(bsFs.bashProfilePath));
         }
-        return console.log(toSourceString(mainPath)); 
+        return console.log(bsFs.toSourceString(mainPath)); 
       },
     }
   },
   open: {
     project: {
       exec: async (flags) => {
-        const projectPath = await getProjectDir(bsBinProject);
+        const projectPath = await bsFs.getProjectDir(bsFs.paths.bin.project);
         const openInCode = Deno.run({
           cmd: ['code', projectPath],
         });
@@ -200,12 +98,12 @@ const commands = {
       const isGlobal = flags.g || flags.global;
       const scripts = flags.s || flags.script;
 
-      const aliasDir = isGlobal ? bsAliasGlobals : await getProjectDir(bsAliasProject);
-      const scriptDir = isGlobal ? bsBinGlobals : await getProjectDir(bsBinProject);
+      const aliasDir = isGlobal ? bsFs.paths.alias.globals : await bsFs.getProjectDir(bsFs.paths.alias.project);
+      const scriptDir = isGlobal ? bsFs.paths.bin.globals : await bsFs.getProjectDir(bsFs.paths.bin.project);
 
       const getScriptsAndAliases = (!alias && !scripts);
       if (getScriptsAndAliases || scripts) {
-        const scriptsNames = await getFileNames(scriptDir, false);
+        const scriptsNames = await bsFs.getFileNames(scriptDir, false);
         scriptsFileNames.push(...scriptsNames);
         Object.assign(result, {
           scripts: scriptsFileNames,
@@ -213,7 +111,7 @@ const commands = {
       }
       if (getScriptsAndAliases || alias) {
         // Map to remove the file type, user do not need to know, they are exec with out.
-        const aliasesNames = (await getFileNames(aliasDir, false)).map((item) => item.split('.')[0]);
+        const aliasesNames = (await bsFs.getFileNames(aliasDir, false)).map((item) => item.split('.')[0]);
         aliasesFileNames.push(...aliasesNames);
         Object.assign(result, {
           aliases: aliasesFileNames,
@@ -223,13 +121,12 @@ const commands = {
       return console.log(JSON.stringify(result, null, 2))
     },
   },
-  r: {
+  run: {
     exec: async (params, flags) => {
-      const projectPath = await getProjectDir(bsBinProject);
-      const filePath = `${projectPath}/${params[0]}`;
-      const doesExist = await exists(filePath);
+      const projectPath = await bsFs.getProjectDir(bsFs.paths.bin.project);
+      const filePath = await bsFs.getScriptPath(params[0], projectPath);
 
-      if (!doesExist) {
+      if (!filePath) {
         return logger.error(`Script: ${params[0]} does not exist in: ${projectPath}`);
       }
 
@@ -258,7 +155,7 @@ function runner(input) {
   const { _, ...rest } = input;
   if (_[0] === 'r' || _[0] === 'run') {
     const params = [..._].slice(1);
-    return commands.r.exec(params, rest);
+    return commands.run.exec(params, rest);
   }
   const func = _.reduce((result, name) => {
    if (!result[name]) {
